Reorder modules
1. Store most of the Python modules inside `src/`
2. Store all the modules related to the gridpack details inside `src/generator`
3. Store all the helper modules (modules to open SSH connections, send emails, ....) inside `src/tools`
4. Retrieve the `mcm_gridpack` module's path dinamically
