+++ src/generator/madgraph_gridpack.py
@@ -1,46 +1,47 @@
 import os
+import glob
 import pathlib
-from gridpack import Gridpack
+from src.gridpack import Gridpack
 
 
-class PowhegGridpack(Gridpack):
+class MadgraphGridpack(Gridpack):
+
+    def prepare_default_card(self):
+        """
+        Copy default cards to local directory
+        """
+        cards_path = self.get_cards_path()
+        job_files_path = self.get_job_files_path()
+        self.logger.debug('Copying %s/*.dat to %s', cards_path, job_files_path)
+        os.system(f'cp {cards_path}/*.dat {job_files_path}')
+        if glob.glob(f'{cards_path}/*_cuts.f'):
+            os.system(f'cp {cards_path}/*_cuts.f {job_files_path}')
 
     def get_run_card(self):
         """
         Get cards from "Template" directory and customize them
-        Get cards from "ModelParams" directory and customize them
-        Glue them together
         """
         dataset_dict = self.get_dataset_dict()
         campaign_dict = self.get_campaign_dict()
         templates_path = self.get_templates_path()
         template_name = dataset_dict['template']
-        template_vars = dataset_dict.get('template_vars', [])
-        template_vars['ebeam1'] = campaign_dict.get('beam', 0)
-        template_vars['ebeam2'] = template_vars['ebeam1']
-        template_vars.update(campaign_dict.get('template_vars', {}))
+        replace_vars = dataset_dict.get('template_vars', [])
+        replace_vars['ebeam1'] = campaign_dict.get('beam', 0)
+        replace_vars['ebeam2'] = replace_vars['ebeam1']
+        replace_vars.update(campaign_dict.get('template_vars', {}))
         input_file_name = os.path.join(templates_path, template_name)
         run_card = self.customize_file(input_file_name,
                                        dataset_dict.get('template_user', []),
-                                       template_vars)
-
-        model_params_path = self.get_model_params_path()
-        model_params_name = dataset_dict['model_params']
-        model_params_vars = dataset_dict.get('model_params_vars', [])
-        model_params_vars.update(campaign_dict.get('model_params_vars', {}))
-        input_file_name = os.path.join(model_params_path, model_params_name)
-        customize_card = self.customize_file(input_file_name,
-                                             dataset_dict.get('model_params_user', []),
-                                             model_params_vars)
-
-        return run_card + '\n' + customize_card
+                                       replace_vars)
+        return run_card
 
     def prepare_run_card(self):
         """
         Get run card and write it to job files dir
         """
         job_files_path = self.get_job_files_path()
-        output_file_name = os.path.join(job_files_path, 'powheg.input')
+        dataset_name = self.data['dataset']
+        output_file_name = os.path.join(job_files_path, f'{dataset_name}_run_card.dat')
         run_card = self.get_run_card()
         self.logger.debug('Writing customized run card %s', output_file_name)
         self.logger.debug(run_card)
@@ -49,19 +50,27 @@ class PowhegGridpack(Gridpack):
 
     def get_customize_card(self):
         """
-        Create card with just the process name for proper Powheg gridpack
-        production 
+        Get cards from "ModelParams" directory and customize them
         """
         dataset_dict = self.get_dataset_dict()
-        template_name = dataset_dict['template']
-        return template_name.split('.', 1)[0]
+        campaign_dict = self.get_campaign_dict()
+        model_params_path = self.get_model_params_path()
+        model_params_name = dataset_dict['model_params']
+        replace_vars = dataset_dict.get('model_params_vars', [])
+        replace_vars.update(campaign_dict.get('model_params_vars', {}))
+        input_file_name = os.path.join(model_params_path, model_params_name)
+        customize_card = self.customize_file(input_file_name,
+                                             dataset_dict.get('model_params_user', []),
+                                             replace_vars)
+        return customize_card
 
     def prepare_customize_card(self):
         """
         Get customize card and write it to job files dir
         """
         job_files_path = self.get_job_files_path()
-        output_file_name = os.path.join(job_files_path, 'process.dat')
+        dataset_name = self.data['dataset']
+        output_file_name = os.path.join(job_files_path, f'{dataset_name}_customizecards.dat')
         customize_card = self.get_customize_card()
         self.logger.debug('Writing customized card %s', output_file_name)
         self.logger.debug(customize_card)
@@ -74,6 +83,7 @@ class PowhegGridpack(Gridpack):
         """
         job_files_path = self.get_job_files_path()
         pathlib.Path(job_files_path).mkdir(parents=True, exist_ok=True)
+        self.prepare_default_card()
         self.prepare_run_card()
         self.prepare_customize_card()
         local_dir = self.local_dir()
